remove chart.js version of galaxy
